requirements.txt
Django==1.11.3
django-bootstrap3==9.0.0
django-rest-framework==0.1.0
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.1
pytz==2017.2
